requirements.txt
numpy~=1.23.5
Pillow~=9.5.0
joblib~=1.2.0
scikit-learn~=1.2.2
scikit-image~=0.20.0
numba~=0.57.1
requests~=2.31.0
scipy~=1.10.1
opencv-python~=4.7.0.72
opencv-python-headless~=4.7.0.72